requirements.txt
# Core dependencies
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Text processing dependencies
nltk>=3.8.1
scikit-learn>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
urllib3>=2.0.0

# Persian text processing
persian-tools>=0.0.8 